problem read db solved

+++ DA-Android_P7_Kotlin-main/app/src/main/java/com/openclassrooms/arista/data/entity/SleepDto.kt
@@ -24,5 +24,5 @@ data class SleepDto (
     var quality: Int,
 
     @ColumnInfo(name = "user_id")
-    var userId: Int //foreign key to UserDto
+    var userId: Long //foreign key to UserDto
 )
